widen raw_temp var to 32bits, adjust reject value to 4050+

+++ Marlin/temperature.h
@@ -225,7 +225,7 @@ class Temperature {
       static millis_t next_bed_check_ms;
     #endif
 
-    static uint16_t raw_temp_value[MAX_EXTRUDERS],
+    static uint32_t raw_temp_value[MAX_EXTRUDERS],
                     raw_temp_bed_value;
 
     // Init min and max temp with extreme values to prevent false errors during startup
